+++ start.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 
-echo "Starting kippo in background..."
-twistd -y kippo.tac -l log/kippo.log --pidfile kippo.pid
+echo "Starting cowrie in background..."
+twistd -y cowrie.tac -l log/cowrie.log --pidfile cowrie.pid